requirements.txt
numpy>=1.26.0,<2.0.0
opencv-python>=4.8.0,<5.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
deepgram-sdk>=3.0.0
ultralytics>=8.0.0
scipy>=1.11.0
filterpy>=1.4.5
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0